Action system. Actions can be undone. @SatanicSanta would approve.
I've been in the car so this is a big commit and also not actually
tested yet. Some random cleanup also done.

An action is one operation, like an edit or deletion. When an action is
performed it's recorded to the ActionHandler and given an ID. We can see
the actions done with !listactionsdone and undo (currently just) one
action with !undoactionbyid. Actions are only saved to the current
runtime though